Rename the class to InterfaceValidator and TestInterfaceValidator , refactor methods and added comments
