+++ .idea/dictionaries/Brayan_Rosas.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Brayan Rosas">
+    <words>
+      <w>paremeters</w>
+    </words>
+  </dictionary>
+</component>
